Refactored package structure of replay tests.
